infra/warehouse/requirements.txt
Faker==35.2.0
future==1.0.0
kafka-python-ng==2.2.3
numpy==2.2.2
pandas==2.2.3
psycopg==3.2.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12. 
tzdata==2025.1
